requirements.txt
hydra-core==1.3.2
invoke==2.2.0
matplotlib==3.10.7
medmnist==3.0.2
numpy==2.3.4
omegaconf==2.3.0
scikit_learn==1.7.2
torch==2.5.1
torchvision==0.20.1
umap==0.1.1
